feat: add a button to buy products
