+++ src/components/Sidebar/index.tsx
@@ -3,7 +3,7 @@ import classNames from 'classnames'
 
 const styles = {
   sidebar:
-    'fixed top-0 right-0 h-screen w-96 overflow-y-auto z-[9999] px-8 pt-12 bg-gray900 block transition ease-in-out duration-[1000ms] h-full overflow-hidden',
+    'fixed top-0 right-0 h-screen w-96 overflow-y-auto z-[9999] px-8 pt-12 bg-gray900 block transition ease-in-out duration-[1000ms] h-full max-h-full overflow-hidden box-border',
   sidebarOpen: 'translate-x-0',
   sidebarClose: 'translate-x-full',
   btnClose: 'absolute top-0 right-0 pt-10 pr-10',
