+++ src/helpers/cart.ts
@@ -1,3 +1,3 @@
-export const countCartItems = (cartItems: app.CartProduct): number => {
+export const countCartItems = (cartItems: app.cartItem): number => {
   return Object.keys(cartItems).length
 }
