+++ src/store/slices/products.ts
@@ -27,6 +27,34 @@ const productSlice = createSlice({
       price: 11,
       stock: 9,
     },
+    {
+      id: 5,
+      name: 'leg warmers',
+      image: 'https://source.unsplash.com/random?product=1',
+      price: 49.9,
+      stock: 10,
+    },
+    {
+      id: 6,
+      name: 'controller',
+      image: 'https://source.unsplash.com/random?product=2',
+      price: 99.9,
+      stock: 1,
+    },
+    {
+      id: 7,
+      name: 'towel',
+      image: 'https://source.unsplash.com/random?product=3',
+      price: 10,
+      stock: 4,
+    },
+    {
+      id: 8,
+      name: 'soda can',
+      image: 'https://source.unsplash.com/random?product=4',
+      price: 11,
+      stock: 9,
+    },
   ],
   reducers: {
     add: (state, action) => {
